+++ test/Types.test.ts
@@ -19,7 +19,7 @@ describe('Public TypeScript types', () => {
 async function runTsBuild(tsVersion: string): Promise<number> {
     const repoRoot = path.join(__dirname, '..', '..');
     const tscPath = path.join(repoRoot, 'node_modules', `typescript${tsVersion}`, 'bin', 'tsc');
-    const projectFile = path.join(repoRoot, 'types', 'tsconfig.json');
+    const projectFile = path.join(repoRoot, 'test', 'types', 'tsconfig.json');
     return new Promise<number>((resolve, reject) => {
         const cmd = cp.spawn('node', [tscPath, '--project', projectFile]);
         cmd.stdout.on('data', function (data) {
