--- types/tsconfig.json
@@ -1,8 +1,0 @@
-{
-    "compilerOptions": {
-        "module": "commonjs",
-        "target": "es6",
-        "strict": true,
-        "outDir": "dist"
-    }
-}
